Removed api route for get user servers and called api directly using SSR
